De-Ankify project (#3)
This moves the project in a direction I'd thought of last time I was working on it... maybe it'd be best to use my own data models & database natively and save `*.apkg` files for import and export.

Now that I want to get good at SwiftData, this makes even more sense; this app will be a great SwiftData playground. So, that's what I'm doing here. The Anki-specific data structures have been moved to a separate module, and I'm creating cleaned up clones of the models for my native storage in SwiftData.

One thing that will happen as part of this is less reliance on HTML/CSS for creating cards. Right now I'm just doing plain text, but I'll probably expand this to Markdown, images, and sound.